+++ sdk/python/pulumi_gcp/container/node_pool.py
@@ -39,6 +39,8 @@ class NodePool(pulumi.CustomResource):
     ) The maximum number of pods per node in this node pool.
     Note that this does not work on node pools which are "route-based" - that is, node
     pools belonging to clusters that do not have IP Aliasing enabled.
+    See the [official documentation](https://cloud.google.com/kubernetes-engine/docs/how-to/flexible-pod-cidr)
+    for more information.
     """
     name: pulumi.Output[str]
     """
@@ -100,6 +102,8 @@ class NodePool(pulumi.CustomResource):
         :param pulumi.Input[float] max_pods_per_node: ) The maximum number of pods per node in this node pool.
                Note that this does not work on node pools which are "route-based" - that is, node
                pools belonging to clusters that do not have IP Aliasing enabled.
+               See the [official documentation](https://cloud.google.com/kubernetes-engine/docs/how-to/flexible-pod-cidr)
+               for more information.
         :param pulumi.Input[str] name: The name of the node pool. If left blank, Terraform will
                auto-generate a unique name.
         :param pulumi.Input[dict] node_config: The node configuration of the pool. See